make LSTM reproducible
